Relocated Combinatorics classes to cross-math. Refactored and optimized old DTW alignment. Created integration tests for Bipace, Cemapp etc.

+++ maven/maltcms/trunk/maltcms-commands/src/main/java/maltcms/commands/distances/IRecurrence.java
@@ -34,7 +34,7 @@ import cross.IConfigurable;
 public interface IRecurrence extends IConfigurable {
 
     public abstract double eval(int row, int column,
-            IArrayD2Double cumDistMatrix, double dij, int[][] predecessors);
+            IArrayD2Double cumDistMatrix, double dij, byte[][] predecessors);
 
     public abstract double eval(int row, int column,
             IArrayD2Double previousRow, IArrayD2Double currentRow, double dij);
